0013 - Correção arquivo RT

+++ Código Fonte/VERSAO-ATUAL/src/Credenciado/Credenciado.Model.Extensoes/Caracterizacoes/ModuloBarragemDispensaLicensa/Data/BarragemDispensaLicencaInternoDa.cs
@@ -90,7 +90,7 @@ namespace Tecnomapas.EtramiteX.Credenciado.Model.Extensoes.Caracterizacoes.Modul
                         caracterizacao.vazaoEnchente = reader.GetValue<decimal>("vazao_enchente");
                         caracterizacao.areaBaciaContribuicao = reader.GetValue<decimal>("area_bacia_contribuicao");
                         caracterizacao.precipitacao = reader.GetValue<decimal>("precipitacao");
-                        caracterizacao.periodoRetorno = reader.GetValue<int>("periodo_retorno");
+                        caracterizacao.periodoRetorno = reader.GetValue<decimal>("periodo_retorno");
                         caracterizacao.coeficienteEscoamento = reader.GetValue<decimal>("coeficiente_escoamento");
                         caracterizacao.tempoConcentracao = reader.GetValue<decimal>("tempo_concentracao");
                         caracterizacao.EquacaoCalculo = reader.GetValue<string>("equacao_calculo");
